The directory was reorganized. Validators to the parameters were added and other minor changes
