+++ backend/controllers/workspace/management/delete-workspace.js
@@ -1,7 +1,9 @@
-const { body } = require('express-validator');
+const { body, param } = require('express-validator');
+const { isUUID } = require('../../../utils/custom-validators');
 const jwt = require('jsonwebtoken');
 const argon2 = require('argon2');
-const { User, Membership, Workspace } = require('../../models/index');
+const checkWorkspaceOwnership = require('../../../utils/check-workspace-ownership');
+const { User, Workspace } = require('../../../models/index');
 
 async function controller(req, res) {
     // Get the needed fields from the body
@@ -14,7 +16,6 @@ async function controller(req, res) {
     try {
         // Get the password of the user
         const { password } = await User.findByPk(uuid, { attributes: ['password'] });
-
         // Verify if the password entered is the correct one in the DB
         if (!(await argon2.verify(password, raw_pswd))) {
             return res.status(403).send('Incorrect password');
@@ -22,35 +23,34 @@ async function controller(req, res) {
     } catch (e) {
         return res.status(500).send('An error has occurred, please try again');
     }
-    let role;
-    try {
-        // Grab the role of the user
-        ({ role } = await Membership.findOne({
-            where: { userId: uuid, workspaceId },
-            attributes: ['role'],
-        }));
-    } catch (e) {
-        return res.status(500).send('An error has occurred, please try again');
-    }
 
     try {
-        //if the role of the user is owner, delete.        
-        if (role === 'owner') {
-            await Workspace.destroy({ where: { id: workspaceId } });
-        } else {
-            return res.status(401).send('You are not the owner of this Workspace.');
+        // If the user is not the owner of that workspace, return
+        if (!(await checkWorkspaceOwnership(uuid, workspaceId))) {
+            return res.status(401).send('You don\'t have permissions to perform that action');
         }
     } catch (e) {
         return res.status(500).send('An error has occurred, please try again');
     }
 
-    res.status(200).send('Workspace Successfully deleted.');
+    try {
+        await Workspace.destroy({ where: { id: workspaceId } });
+    } catch (e) {
+        return res.status(500).send('An error has occurred, please try again');
+    }
+
+    res.status(200).send('Workspace successfully deleted');
 }
 
 const validators = [
     body('password')
         .isLength({ max: 50 })
-        .withMessage('Must be maximum 50 characters long')
+        .withMessage('Must be maximum 50 characters long'),
+    param('workspaceId')
+        .isLength({ min: 36, max: 36 })
+        .withMessage('Invalid workspace ID')
+        .custom(isUUID)
+        .withMessage('Invalid workspace ID'),
 ];
 
 module.exports = {
